# docs/swagger.yaml
definitions:
  twitch.ChatMessage:
    properties:
      channel:
        type: string
      created:
        type: string
      id:
        type: string
      message:
        type: string
      originalTime:
        type: string
      raw:
        type: string
      user:
        $ref: '#/definitions/twitch.ChatUser'
    type: object
  twitch.ChatUser:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  twitch.TushqaQuote:
    properties:
      channel:
        type: string
      created:
        type: string
      id:
        type: string
      message:
        type: string
    type: object
  web.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  web.Rss:
    properties:
      channel:
        $ref: '#/definitions/web.RssChannel'
      version:
        type: string
      xmlname:
        type: object
    type: object
  web.RssChannel:
    properties:
      items:
        items:
          $ref: '#/definitions/web.RssChannelItem'
        type: array
      lastBuildDate:
        type: string
      link:
        type: string
      title:
        type: string
    type: object
  web.RssChannelItem:
    properties:
      description:
        type: string
      link:
        type: string
      originalDate:
        type: string
      originalUrl:
        type: string
      pubDate:
        type: string
      title:
        type: string
      uid:
        type: string
    type: object
info:
  contact: {}
paths:
  /dl/{fileId}:
    get:
      parameters:
      - description: File id
        in: path
        name: fileId
        required: true
        type: string
      produces:
      - application/octet-stream
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      tags:
      - File controller
  /kinozal/rss:
    get:
      produces:
      - text/xml
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Rss'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      tags:
      - Kinozal controller
  /lostfilm/rss:
    get:
      parameters:
      - description: Quality filter
        in: query
        name: quality
        type: string
      produces:
      - text/xml
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Rss'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      tags:
      - LostFilm controller
  /proxy:
    get:
      parameters:
      - description: Url for proxied GET request
        in: query
        name: url
        required: true
        type: string
      - description: override content-type header
        in: query
        name: responseHeaderContentType
        type: string
      produces:
      - text/plain
      responses: {}
      tags:
      - Proxy controller
  /twitch/messages:
    get:
      parameters:
      - description: Channel filter
        in: query
        name: channel
        type: string
      - description: Message list limit
        in: query
        maximum: 100
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/twitch.ChatMessage'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      tags:
      - Twitch controller
  /twitch/tushqa:
    get:
      parameters:
      - description: Quotes limit
        in: query
        maximum: 100
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/twitch.TushqaQuote'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      tags:
      - Twitch controller
swagger: "2.0"